# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CutAI - AI Video Editor API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{id}/videos:
    get:
      operationId: listProjectVideos
      tags: [videos]
      summary: List videos in a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of videos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Video"
  /videos:
    post:
      operationId: uploadVideo
      tags: [videos]
      summary: Upload a video file (multipart)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UploadVideoBody"
      responses:
        "201":
          description: Uploaded video
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
  /videos/{id}/thumbnail.jpg:
    get:
      operationId: getVideoThumbnail
      tags: [videos]
      summary: Get a thumbnail frame from a video
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: t
          in: query
          required: false
          schema:
            type: number
            description: Time in seconds to extract thumbnail (default 1.0)
      responses:
        "200":
          description: JPEG thumbnail image
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        "404":
          description: Video not found
  /videos/{id}/stream:
    get:
      operationId: streamVideo
      tags: [videos]
      summary: Stream a video file with range support
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Video file stream
          content:
            video/mp4:
              schema:
                type: string
                format: binary
        "206":
          description: Partial content (range request)
  /videos/{id}:
    get:
      operationId: getVideo
      tags: [videos]
      summary: Get video by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Video details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
        "404":
          description: Not found
    delete:
      operationId: deleteVideo
      tags: [videos]
      summary: Delete video
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Video:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        filename:
          type: string
        originalName:
          type: string
        mimeType:
          type: string
        sizeBytes:
          type: integer
        durationSeconds:
          type: ["number", "null"]
        width:
          type: ["integer", "null"]
        height:
          type: ["integer", "null"]
        fps:
          type: ["number", "null"]
        filePath:
          type: string
        status:
          type: string
          enum: [uploading, ready, processing, error, transcoding]
        transcript:
          type: ["string", "null"]
        sceneAnalysis:
          type: ["string", "null"]
        beatData:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - projectId
        - filename
        - originalName
        - mimeType
        - sizeBytes
        - filePath
        - status
        - createdAt
    UploadVideoBody:
      type: object
      properties:
        projectId:
          type: string
        file:
          type: string
          format: binary
      required:
        - projectId
        - file
...